requirements.txt
Flask
python-dotenv
openai
requests
skyfield
